Remove mention of tests from qualifier and add examples
I have removed the mention of "should pass all tests" from the readme
since we no longer plan to give those to the participants. Instead, I
have added some examples that illustrate the desired behavior.

In addition, I've included some strong words about leaving the
function signature as it is, since that's important for the automated
test suite we plan on using to judge the entries. I've also included
information about the deadline and a placeholder for the link to the
submission form.