Clarify decimal seconds and truncated format
We received a lot of questions about the number of decimal seconds
the function should support; I've added the clarification of "up to
six decimals".

Another question that came up frequently was if a mixture of regular
and truncated representations was allowed. Since the standard seems
to disallow that, but our requirements did not, I've added a note
that says that we're only going to test the function with strings in
which both parts are in the same representation, but that we're not
going to take points away if the function accepts strings with a
mixed representation.